bugfix: filter number of cores properly
This code was written on a four-core machine, and so I filtered the -4
extensions to each benchmark without realizing what it represented. This
caused it to break when run on machines with a different core count than
mine. The breakage resulted in each benchmark being given the default
color of white, instead of the preset colors.

Signed-off-by: Chris Waldon <chris.waldon@ibm.com>